requirements.txt
Django
Pillow
Celery
djangorestframework
drf-yasg